pleromanna/requirements.txt
zappa
zappa-django-utils
wagtail
wagtailmenus
wagtail-condensedinlinepanel
django_storages
psycopg2-binary
